refactored the feature tests; each manage test is now split into its CRUD features, with the miscellaneous tests moved to the appropriate subfeature
